refactoring to take socket and types as Provider props
