+++ package.json
@@ -13,6 +13,7 @@
     "@polkadot/util": "^1.2.1",
     "@polkadot/util-crypto": "^1.2.1",
     "gh-pages": "^2.1.1",
+    "prop-types": "^15.7.2",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-scripts": "3.1.1",
