+++ src/substrate/useSubstrate.js
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useCallback } from 'react';
 import { ApiPromise, WsProvider } from "@polkadot/api";
 
-import { SubstrateContext } from '../contexts/SubstrateContext';
+import { SubstrateContext } from './SubstrateContext';
 
 const useSubstrate = () => {
   let [state, dispatch] = useContext(SubstrateContext);
@@ -9,18 +9,13 @@ const useSubstrate = () => {
   // `useCallback` so that returning memoized function and not created everytime,
   //    and thus re-render.
   const connect = useCallback(async() => {
-    let { api, socket } = state;
+    let { api, socket, types } = state;
     if (api) return;
 
     const provider = new WsProvider(socket);
 
-    // More information on custom types
-    // https://github.com/polkadot-js/apps/blob/master/packages/app-settings/src/md/basics.md
-    //const TYPES = {"MyNumber": "u32"};
-    const TYPES = {};
-
     try {
-      api = await ApiPromise.create({provider, types: TYPES});
+      api = await ApiPromise.create({ provider, types });
       dispatch({ type: 'CONNECT', payload: api });
       api.isReady.then(() => dispatch({type: 'CONNECT_SUCCESS'}));
 
@@ -34,7 +29,7 @@ const useSubstrate = () => {
     connect();
   }, [connect]);
 
-  return { state, dispatch };
+  return { ...state, dispatch };
 }
 
 export default useSubstrate;
